+++ src/react/components/error-legend.jsx
@@ -1,0 +1,16 @@
+/**
+ * Display error legend
+ */
+wplv.ErrorLegend = React.createClass({
+
+	// Render component
+	render: function() {
+		return (
+			<ul className="error-indicator-legend">
+				<li className="php-fatal-error">Fatal Error</li>
+				<li className="php-warning">PHP Warning</li>
+				<li className="php-notice">PHP Notice</li>
+			</ul>
+		);	
+	}
+});